move rendering logic into Column component
